# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIMULA Backend",
    "description": "API PMR Mula. Router → service → SQLAlchemy ORM → PostgreSQL.",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/modules": {
      "get": {
        "tags": [
          "Modules"
        ],
        "summary": "Modules",
        "operationId": "modules_api_v1_modules_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleSummary"
                  },
                  "title": "Response Modules Api V1 Modules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Modules"
        ],
        "summary": "Module Create",
        "operationId": "module_create_api_v1_modules_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/{mid}": {
      "put": {
        "tags": [
          "Modules"
        ],
        "summary": "Module Update",
        "operationId": "module_update_api_v1_modules__mid__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Modules"
        ],
        "summary": "Module Detail",
        "operationId": "module_detail_api_v1_modules__mid__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/{mid}/submit-review": {
      "post": {
        "tags": [
          "Validation"
        ],
        "summary": "Submit Review",
        "operationId": "submit_review_api_v1_modules__mid__submit_review_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/{mid}/review": {
      "post": {
        "tags": [
          "Validation"
        ],
        "summary": "Review Module",
        "operationId": "review_module_api_v1_modules__mid__review_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/{mid}/lessons": {
      "post": {
        "tags": [
          "Lessons"
        ],
        "summary": "Lesson Create",
        "operationId": "lesson_create_api_v1_modules__mid__lessons_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lessons/{lid}": {
      "put": {
        "tags": [
          "Lessons"
        ],
        "summary": "Lesson Update",
        "operationId": "lesson_update_api_v1_lessons__lid__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Lessons"
        ],
        "summary": "Lesson Delete",
        "operationId": "lesson_delete_api_v1_lessons__lid__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lessons/{lid}/complete": {
      "post": {
        "tags": [
          "Learning"
        ],
        "summary": "Complete",
        "operationId": "complete_api_v1_lessons__lid__complete_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learning/progress": {
      "get": {
        "tags": [
          "Learning"
        ],
        "summary": "Progress",
        "operationId": "progress_api_v1_learning_progress_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProgressOut"
                  },
                  "type": "array",
                  "title": "Response Progress Api V1 Learning Progress Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/learning/attempts": {
      "get": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Attempts",
        "operationId": "attempts_api_v1_learning_attempts_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttemptOut"
                  },
                  "title": "Response Attempts Api V1 Learning Attempts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttemptOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "quiz_id": {
            "type": "integer",
            "title": "Quiz Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "answers": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Answers"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "quiz_id",
          "score",
          "passed",
          "answers",
          "created_at"
        ],
        "title": "AttemptOut"
      },
...
      "CompletionOut": {
        "properties": {
          "completed": {
            "type": "boolean",
            "title": "Completed"
          },
          "xp_awarded": {
            "type": "integer",
            "title": "Xp Awarded"
          }
        },
        "type": "object",
        "required": [
          "completed",
          "xp_awarded"
        ],
        "title": "CompletionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LessonCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 10,
            "title": "Body"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          },
          "position": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body",
          "position"
        ],
        "title": "LessonCreate"
      },
      "LessonOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "module_id",
          "title",
          "body",
          "image_url",
          "video_url",
          "position"
        ],
        "title": "LessonOut"
      },
...
      "ModuleCreate": {
        "properties": {
          "topic_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Topic Id",
            "default": 1
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 3,
            "title": "Description"
          },
          "level": {
            "type": "integer",
            "maximum": 1000.0,
            "exclusiveMinimum": 0.0,
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "level"
        ],
        "title": "ModuleCreate"
      },
      "ModuleDetail": {
        "properties": {
          "topic_id": {
            "type": "integer",
            "title": "Topic Id",
            "default": 1
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonOut"
            },
            "type": "array",
            "title": "Lessons"
          },
          "quizzes": {
            "items": {
              "$ref": "#/components/schemas/QuizOut"
            },
            "type": "array",
            "title": "Quizzes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "level",
          "status",
          "review_note",
          "created_by",
          "reviewed_by",
          "created_at",
          "lessons",
          "quizzes"
        ],
        "title": "ModuleDetail"
      },
      "ModuleOut": {
        "properties": {
          "topic_id": {
            "type": "integer",
            "title": "Topic Id",
            "default": 1
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "level",
          "status",
          "review_note",
          "created_by",
          "reviewed_by",
          "created_at"
        ],
        "title": "ModuleOut"
      },
      "ModuleSummary": {
        "properties": {
          "topic_id": {
            "type": "integer",
            "title": "Topic Id",
            "default": 1
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "level",
          "status",
          "review_note"
        ],
        "title": "ModuleSummary"
      },
...
      "ProgressOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "level",
          "total",
          "completed"
        ],
        "title": "ProgressOut"
      },
...
      "QuizOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "pass_score": {
            "type": "integer",
            "title": "Pass Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "module_id",
          "title",
          "kind",
          "pass_score"
        ],
        "title": "QuizOut"
      },
...
      "Review": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "approved",
          "note"
        ],
        "title": "Review"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}